Add tabs to about section
