+++ app/javascript/app/components/about-description/about-description-component.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+// import styles from './about-description-styles.scss';
+
+const About = () => (
+  <div>
+    Climate Watch is an online platform designed to empower policymakers,
+    researchers, media and other stakeholders with the open climate data,
+    visualizations and resources they need to gather insights on national and
+    global progress on climate change. Climate Watch brings together dozens of
+    datasets for the first time to let users analyze and compare the Nationally
+    Determined Contributions (NDCs) under the Paris Agreement, access historical
+    emissions data, discover how countries can leverage their climate goals to
+    achieve their sustainable development objectives, and use models to map new
+    pathways to a lower carbon, prosperous future. This free platform enables
+    users to create and share custom data visualizations and comparisons of
+    national climate commitments. It contributes to the goals of the Paris
+    Agreement by using open data to increase transparency and accountability,
+    and provide actionable analysis on how countries can enhance their efforts
+    to combat climate change. Climate Watch includes: Data and visualizations on
+    all countries’ greenhouse gas emissions; A comprehensive, user-friendly
+    database of all countries’ Nationally Determined Contributions (NDCs);
+    Comprehensive mapping of linkages between Nationally Determined
+    Contributions (NDCs) and the Sustainable Development Goals (SDGs) and
+    associated targets of the 2030 Agenda for Sustainable Development. Data and
+    visuals of emissions scenario pathways for major emitting countries, derived
+    from a growing library of models; National and sectoral profile pages that
+    offer a snapshot of climate progress, risks and vulnerabilities; and The
+    ability to download data and create, save and share customized data
+    visualizations through My Climate Watch. Climate Watch is implemented by
+    World Resources Institute in collaboration with partners.
+  </div>
+);
+
+export default About;
