+++ app/javascript/app/components/about-description/about-description-styles.scss
@@ -1,0 +1,2 @@
+@import '~styles/layout.scss';
+@import '~styles/settings.scss';
