+++ app/javascript/app/components/about-description/about-description.js
@@ -1,0 +1,3 @@
+import Component from './about-description-component';
+
+export default Component;
