+++ app/javascript/app/routes.js
@@ -16,6 +16,8 @@ import CountryNdcOverview from 'components/country-ndc-overview';
 import NDCSearch from 'pages/ndc-search';
 import GHGEmissions from 'pages/ghg-emissions';
 import About from 'pages/about';
+import AboutDescription from 'components/about-description';
+import AboutPartners from 'components/about-partners';
 import GHGCountryEmissions from 'components/country-ghg';
 import NDCSDGLinkages from 'components/country-ndc-sdg-linkages';
 import ClimateVulnerability from 'components/country-climate-vulnerability';
@@ -154,10 +156,25 @@ export default [
       {
         path: '/about',
         component: About,
-        exact: true,
         nav: true,
         label: 'ABOUT',
-        headerImage: 'about'
+        headerImage: 'about',
+        routes: [
+          {
+            path: '/about',
+            component: AboutDescription,
+            exact: true,
+            anchor: true,
+            label: 'About Climate Watch'
+          },
+          {
+            path: '/about/partners',
+            component: AboutPartners,
+            exact: true,
+            anchor: true,
+            label: 'Partners'
+          }
+        ]
       },
       {
         path: '/error-page',
